feat: migrate and enhance entity structure, add base entity and new models

+++ src/main/resources/db/migration/V1__init_schema.sql
@@ -144,7 +144,7 @@ CREATE TABLE budget_item
     ferragem_qtd_pontos_amarracao_estribo       INTEGER,
     ferragem_qtd_arame_por_ponto_amarracao      REAL,
     ferragem_qtd_estribos                       INTEGER,
-    ferragem_comprimento_total_estribos         INTEGER,
+    ferragem_comprimento_total_estribos         REAL,
     ferragem_qtd_total_pontos_amarracao_estribo INTEGER,
     quantity                                    REAL,
     ferragem_qtd_total_arame                    REAL,
